# app/api/v1/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Pikcel
    description: A production-ready Go service deployed on Kubernetes
    version: 1.0.0
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /api/v1/canvas:
        get:
            tags:
                - api
            summary: CanvasGet api
            operationId: api#CanvasGet
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Canvas'
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/APICanvasGetUnauthenticatedResponseBody'
                "403":
                    description: Forbidden response.
                    schema:
                        $ref: '#/definitions/APICanvasGetAccessDeniedResponseBody'
            schemes:
                - http
    /api/v1/openapi.json:
        get:
            tags:
                - api
            summary: Download gen/http/openapi3.json
            operationId: api#/api/v1/openapi.json
            responses:
                "200":
                    description: File downloaded
                    schema:
                        type: file
            schemes:
                - http
definitions:
    APICanvasGetAccessDeniedResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: false
        description: CanvasGet_access_denied_Response_Body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    APICanvasGetUnauthenticatedResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: CanvasGet_unauthenticated_Response_Body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    Canvas:
        title: 'Mediatype identifier: application/vnd.pikcel.canvas`; view=default'
        type: object
        properties:
            height:
                type: integer
                example: 1197379349
                format: int32
            id:
                type: string
                example: Quibusdam mollitia.
            width:
                type: integer
                example: 207392891
                format: int32
        description: CanvasGetResponseBody result type (default view)
        example:
            height: 940953074
            id: Nihil tempora beatae nostrum placeat.
            width: 1228455270
        required:
            - id
            - width
            - height